requirements.txt
asttokens==2.0.8
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
fonttools==4.37.1
idna==3.3
install==1.3.5
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mecab-python3==1.0.5
nest-asyncio==1.5.5
numpy==1.23.3
packaging==21.3
pandas==1.4.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
urllib3==1.26.12
wcwidth==0.2.5
wordcloud==1.8.2.2
